+++ Core/Battle/Menu/BattleMenu.cs
@@ -10,7 +10,7 @@ namespace OpenVIII
     /// </summary>
     public partial class BattleMenu : Menu
     {
-        public bool CrisisLevel => ((IGMData_Commands)Data[SectionName.Commands]).CrisisLevel;
+        public bool CrisisLevel => ((IGMData.Commands)Data[SectionName.Commands]).CrisisLevel;
 
         //private Mode _mode = Mode.Waiting;
 
@@ -50,7 +50,7 @@ namespace OpenVIII
             NoInputOnUpdate = true;
             Size = new Vector2 { X = 880, Y = 636 };
             Data.Add(SectionName.HP, new IGMData_HP(new Rectangle((int)(Size.X - 389), 507, 389, 126), Damageable));
-            Data.Add(SectionName.Commands, new IGMData_Commands(new Rectangle(50, (int)(Size.Y - 204), 210, 192), Damageable, true));
+            Data.Add(SectionName.Commands, new IGMData.Commands(new Rectangle(50, (int)(Size.Y - 204), 210, 192), Damageable, true));
             Data.Add(SectionName.Renzokeken, new IGMData_Renzokeken(new Rectangle(0, 500, (int)Size.X, 124)));
             Data.ForEach(x => x.Value.AddModeChangeEvent(ref ModeChangeHandler));
             SetMode(Damageable.BattleMode.ATB_Charging);
